requirements.txt
# Architectural Style Recognition Benchmark
# Python 3.8+

# Core ML
torch>=2.0.0
torchvision>=0.15.0

# Data processing
numpy>=1.24.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# ML utilities
scikit-learn>=1.2.0
scipy>=1.10.0

# Progress bars
tqdm>=4.65.0

# Dataset download (optional)
kagglehub>=0.2.0

# Hybrid image collection (optional, for download_hybrid_images.py)
duckduckgo_search>=4.0.0
requests>=2.28.0

# Cloud training (optional)
modal>=0.60.0

